Update tiny.sh for strace

+++ tiny.sh
@@ -257,10 +257,12 @@ int main(int argc, char *argv[]) {
 	gcc -static -o /tmp/user /tmp/user.c || exit 108
 	chmod +x /tmp/user || exit 109
 	strace -o /tmp/user.txt /tmp/user
-	cat /tmp/user.txt | grep ^openat | awk -F\" '{print $2;}' | egrep "\.so(\.[0-9\.])*$" | while read f; do
-		d=boot$(dirname "$f")
-		sudo mkdir -p "$d"
-		sudo cp -v "$f" "$d" || exit 110
+	cat /tmp/user.txt | grep ^openat | awk -F\" '{print $2;}' | egrep "\.so(\.[0-9\.])*$" | sort | uniq | while read f; do
+		if [ -f "$f" ]; then
+			d=boot$(dirname "$f")
+			sudo mkdir -p "$d"
+			sudo cp -v "$f" "$d" || exit 110
+		fi
 	done || exit 111
 
 	sudo chown -R root.root boot/etc || 250
